:white_check_mark: test(update): Member 관련 test 분리 및 추가
